Backend/requirements.txt
setuptools>=70.0.0
wheel>=0.41.0
anyio>=3.7.1,<4.0.0
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
httpx>=0.24,<0.28
requests==2.32.3
shapely>=2.0; python_version >= '3.9'
browser-use==0.1.0
pydantic>=2.7.0
supabase>=2.6.0
python-dotenv==1.0.1
websockets>=13.0
